refactor: extract cleanup and list commands, rename edit function
