+++ src/command/edit.rs
@@ -11,19 +11,22 @@ pub enum Error {
     Exit(ExitStatus),
 }
 
-pub(crate) fn edit(path: &Path) -> Result<(), Error> {
+pub(crate) fn run_edit(path: &Path) -> Result<(), super::Error> {
     let _span = tracing::trace_span!("edit", ?path).entered();
 
     let mut cmd = if atty::is(Stream::Stdout) && atty::is(Stream::Stderr) && atty::is(Stream::Stdin)
     {
-        open_cmd::open_editor(path.to_owned())?
+        open_cmd::open_editor(path.to_owned()).map_err(Error::Start)?
     } else {
-        open_cmd::open(path.to_owned())?
+        open_cmd::open(path.to_owned()).map_err(Error::Start)?
     };
 
-    let status = cmd.status().map_err(open_cmd::Error::from)?;
+    let status = cmd.status()
+        .map_err(open_cmd::Error::from)
+        .map_err(Error::Start)
+        .map_err(super::Error::Edit)?;
     if !status.success() {
-        return Err(Error::Exit(status));
+        return Err(super::Error::Edit(Error::Exit(status)));
     }
 
     Ok(())
