+++ src/command/mod.rs
@@ -1,14 +1,18 @@
 //! See [`Command`].
 
+mod cleanup;
 mod diff;
 mod edit;
+mod list;
 mod status;
 
 use structopt::StructOpt;
 use thiserror::Error;
 
-pub(crate) use edit::{edit, Error as EditError};
+pub(crate) use cleanup::run_cleanup;
+pub(crate) use edit::run_edit;
 pub(crate) use diff::run_diff;
+pub(crate) use list::run_list;
 pub(crate) use status::run_status;
 
 /// Errors that can occur while running commands.
@@ -17,11 +21,23 @@ pub enum Error {
     /// Error occurred while printing the help message.
     #[error("error while printing help message: {0}")]
     PrintHelp(#[from] structopt::clap::Error),
+    /// An error occurred while running the cleanup command.
+    #[error("error after cleaning up {success_count} log files: {error}")]
+    Cleanup {
+        /// The number of files successfully cleaned.
+        success_count: u32,
+        /// The error that occurred.
+        #[source]
+        error: crate::checkers::history::operation::Error,
+    },
     /// Error occurred while running the diff command.
-    #[error("error while running hoard diff")]
+    #[error("error while running hoard diff: {0}")]
     Diff(#[source] crate::hoard::iter::Error),
+    /// Error occurred while running the edit command.
+    #[error("error while running hoard edit: {0}")]
+    Edit(#[from] edit::Error),
     /// Error occurred while running the status command.
-    #[error("error while running hoard status")]
+    #[error("error while running hoard status: {0}")]
     Status(#[source] crate::hoard::iter::Error),
 }
